distsql: move SelectResult to file "select_result.go" (#7054)
